requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.9
MarkupSafe==3.0.3
python-frontmatter==1.1.0
PyYAML==6.0.3
Werkzeug==3.1.3
